refactor(cockpit): split review packet renderer
